사진 안뜨는거 이슈 해결

+++ src/main/java/com/green/ca2sa/cafe/CafeService.java
@@ -42,7 +42,7 @@ public class CafeService {
 
     // 카페 정보 변경
     @Transactional
-    public int updCafe(MultipartFile pic, CafePutReq p){
+    public int updCafe(MultipartFile pic, CafePutReq p) {
         if(pic == null){
             return cafeMapper.updCafe(p);
         }
@@ -50,7 +50,11 @@ public class CafeService {
         long cafeId = p.getCafeId();
 
         String folderPath = String.format("cafe/%d",cafeId);
-        myFileUtils.deleteFile(folderPath);
+
+        if(myFileUtils.existFile(folderPath)){
+            myFileUtils.deleteFile(folderPath);
+        }
+        myFileUtils.makeFolders(folderPath);
 
         String fileName = myFileUtils.makeRandomFileName(pic);
         p.setCafePic(fileName);
